Facilitate bundling of sharp
Dynamic require statements make it more difficult for a bundler to figure out the needed modules. In this case there's not much benefit in code reduction either, so just revert to static strings inside the require statement.

This (together with providing a list of dlls to include) makes it possible to bundle a nodejs application using sharp into a stand-alone executable using nexe.