docs(clone): add promise example
It took me some time to figure out how to properly create a pipeline and know
when it's done, while https://github.com/lovell/sharp/issues/524 provides an
example about listening to the finish event, I needed one for toFile.

Now I think we should also document the example from
https://github.com/lovell/sharp/issues/524 but I'd like you the maintainer to
create it because I am not 100% sure this is the most elegant way.

Thanks for the awesome library!