convert tests that where not passing in jest to the default expect pattern
Some of the tests where not passing using the node assert library.

Over time we could normalize all tests to use the newer expect syntax which has a lot of nice assertions which are missing from the node assert lib.

But the changes in this commit was the minimum to get all tests passing.