Issue template: ask for complete error message
The error messages generated by sharp are designed to provide
people with as much information about the situation as possible.
If someone is asked to copy/paste the error, it's more likely
they might actually read it, which might then allow them to help
themselves without the need to open a new issue.